+++ crates/biome_js_formatter/src/js/statements/empty_statement.rs
@@ -1,5 +1,5 @@
 use crate::prelude::*;
-use biome_formatter::write;
+use biome_formatter::{write, CstFormatContext};
 
 use biome_js_syntax::{JsEmptyStatement, JsEmptyStatementFields, JsSyntaxKind};
 use biome_rowan::{AstNode, SyntaxNodeOptionExt};
@@ -11,6 +11,18 @@ impl FormatNodeRule<JsEmptyStatement> for FormatJsEmptyStatement {
     fn fmt_fields(&self, node: &JsEmptyStatement, f: &mut JsFormatter) -> FormatResult<()> {
         let JsEmptyStatementFields { semicolon_token } = node.as_fields();
         let parent_kind = node.syntax().parent().kind();
+
+        let leading_comments_with_break = f
+            .context()
+            .comments()
+            .leading_comments(node.syntax())
+            .iter()
+            .any(|comment| comment.lines_before() > 0 || comment.kind().is_line());
+
+        if leading_comments_with_break {
+            write!(f, [hard_line_break()])?;
+        }
+        write!(f, [format_leading_comments(node.syntax())])?;
 
         if matches!(
             parent_kind,
@@ -29,5 +41,10 @@ impl FormatNodeRule<JsEmptyStatement> for FormatJsEmptyStatement {
         } else {
             write!(f, [format_removed(&semicolon_token?)])
         }
+    }
+
+    fn fmt_leading_comments(&self, _: &JsEmptyStatement, _: &mut JsFormatter) -> FormatResult<()> {
+        // Formatted inside of `fmt_fields`
+        Ok(())
     }
 }