+++ crates/biome_js_formatter/src/ts/lists/tuple_type_element_list.rs
@@ -9,7 +9,7 @@ impl FormatRule<TsTupleTypeElementList> for FormatTsTupleTypeElementList {
     type Context = JsFormatContext;
 
     fn fmt(&self, node: &TsTupleTypeElementList, f: &mut JsFormatter) -> FormatResult<()> {
-        let trailing_separator = FormatTrailingComma::All.trailing_separator(f.options());
+        let trailing_separator = FormatTrailingComma::ES5.trailing_separator(f.options());
 
         f.join_with(&soft_line_break_or_space())
             .entries(