+++ crates/biome_service/src/configuration/parse/json/javascript/formatter.rs
@@ -34,10 +34,13 @@ impl DeserializationVisitor for JavascriptFormatterVisitor {
             "trailingComma",
             "semicolons",
             "arrowParentheses",
+            "bracketSpacing",
+            "bracketSameLine",
             "enabled",
             "indentStyle",
             "indentSize",
             "indentWidth",
+            "lineEnding",
             "lineWidth",
         ];
         let mut result = Self::Output::default();
@@ -70,6 +73,14 @@ impl DeserializationVisitor for JavascriptFormatterVisitor {
                     result.arrow_parentheses =
                         Deserializable::deserialize(&value, &key_text, diagnostics);
                 }
+                "bracketSpacing" => {
+                    result.bracket_spacing =
+                        Deserializable::deserialize(&value, &key_text, diagnostics);
+                }
+                "bracketSameLine" => {
+                    result.bracket_same_line =
+                        Deserializable::deserialize(&value, &key_text, diagnostics);
+                }
                 "enabled" => {
                     result.enabled = Deserializable::deserialize(&value, &key_text, diagnostics);
                 }
@@ -88,6 +99,10 @@ impl DeserializationVisitor for JavascriptFormatterVisitor {
                 }
                 "indentWidth" => {
                     result.indent_width =
+                        Deserializable::deserialize(&value, &key_text, diagnostics);
+                }
+                "lineEnding" => {
+                    result.line_ending =
                         Deserializable::deserialize(&value, &key_text, diagnostics);
                 }
                 "lineWidth" => {